refactor: renew feed
