+++ src/assets/style/Main.scss
@@ -3,7 +3,7 @@
 .main-feed-container {
   width: 100%;
   height: 30vh;
-  max-height: 280px;
+  max-height: 220px;
 
   display: flex;
   flex-direction: column;
@@ -17,12 +17,11 @@
 
 .main-feed-title-container {
   width: calc(100% - 64px);
-  height: 10%;
+  height: 40px;
 
   display: flex;
   justify-content: space-between;
-  align-items: flex-start;
-  flex-grow: 1;
+  align-items: center;
 }
 
 .main-feed-content-container {
@@ -36,7 +35,7 @@
   margin: $margin-base;
 
   width: $margin-width;
-  height: 140px;
+  height: 110px;
 
   border-radius: $box-border-radius;
