+++ src/components/Presenter/Main/Feed/FeedContainer.tsx
@@ -2,8 +2,6 @@
  * @description main 화면의 상단 피드
  */
 
-import React, { useState } from 'react';
-
 import FeedTitle from './FeedTitle';
 import FeedContent from './FeedContent';
 
@@ -11,7 +9,15 @@ export default function Feed() {
   return (
     <div className={'main-feed-container'}>
       <FeedTitle title={'나의 피드'} />
-      <FeedContent />
+      <FeedContent
+        feedList={[
+          {
+            title: '오늘의 강의',
+            id: 'lecture',
+          },
+          { title: '학사 일정', id: 'calender' },
+        ]}
+      />
     </div>
   );
 }
