+++ src/components/Presenter/Main/Feed/Lecture/type.ts
