+++ src/components/Presenter/Main/Feed/type.ts
@@ -1,1 +1,6 @@
 export type TabType = 'lecture' | 'calender';
+
+export interface feedListState {
+  title: string;
+  id: TabType;
+}
